requirements.txt
djangorestframework 
Django
requests
beautifulsoup4
black